Switched const static int members to actual type-safe enums and casts

+++ eva1/bead1/LovagiTorna/widget.cpp
@@ -11,8 +11,8 @@ Widget::Widget(QWidget *parent) :
     for(int i = 0; i < 64; ++i) {
         QBoardButton* btn = new QBoardButton();
         //btn->setText((i%2==0)!=((i/8)%2==0)?QString("FEKETE"):QString("feher"));
-        btn->setColor((i%2==0)!=((i/8)%2==0) ? Game::BoardColor::Black : Game::BoardColor::White);
-        btn->setPiece(Game::PieceType::BlackLeft);
+        btn->setColor((i%2==0)!=((i/8)%2==0) ? static_cast<int>(Game::BoardColor::Black) : static_cast<int>(Game::BoardColor::White));
+        btn->setPiece(static_cast<int>(Game::PieceType::BlackLeft));
         gl->addWidget(btn, i/8, i%8);
         btnVector.push_back(btn);
     }
